test: split api.test.js to match module structure
- Removed packages/github/src/api.test.js
- Created separate test files for each API module:
  - api/types.test.js - tests for isPullRequest type guard
  - api/repo-info.test.js - tests for getGitHubRepoInfo
  - api/pull-request.test.js - tests for fetchPullRequest
  - api/issue.test.js - tests for fetchIssue
